Documentation, light refactor and demo notebooks
