+++ .vscode/pydistsim.code-workspace
@@ -16,6 +16,6 @@
     "remote.autoForwardPortsFallback": 0,
     "remote.autoForwardPorts": false,
     "cSpell.enabled": true,
-    "cSpell.words": ["Santoro"]
+    "cSpell.words": ["Santoro", "yoyo"]
   }
 }
