+++ docs/reference/networkmixin.rst
@@ -31,7 +31,7 @@ Communication properties
 ************************
 .. currentmodule:: pydistsim.network.behavior
 
-The communication properties of the network are defined in the :class:`behaviorModel` class. This
+The communication properties of the network are defined in the :class:`NetworkBehaviorModel` class. This
 class defines the three parameters that are used to model the communication properties of the network:
 
    #. Whether or not the messages being sent at a given time will arrive in the same order they were sent (massage
@@ -40,7 +40,7 @@ class defines the three parameters that are used to model the communication prop
    #. The frequency at which the messages will be lost.
 
 To apply these properties to a network, simply set :attr:`NetworkMixin.behavioral_properties` to an instance of
-:class:`behaviorModel`.
+:class:`NetworkBehaviorModel`.
 
 
 Some predefined models are available in the :class:`ExampleProperties` class:
