--- docs/reference/nodealgorithm.rst
@@ -1,8 +1,0 @@
-.. _nodealgorithm:
-.. currentmodule:: pydistsim.algorithm.node_algorithm
-
-################################################
-Implementation details of :class:`NodeAlgorithm`
-################################################
-
-TODO RST
