--- docs/reference/simulation.rst
@@ -1,8 +1,0 @@
-.. _refsimulation:
-.. currentmodule:: pydistsim.simulation
-
-################################################
-Implementation details of :class:`Simulation`
-################################################
-
-TODO RST
