+++ docs/tutorials/index.rst
@@ -4,7 +4,7 @@
 Tutorials
 #########
 
-Tutorials assume that the PyDistrSim and all required packages are installed. If not, please refer to
+Tutorials assume that the PyDistSim and all required packages are installed. If not, please refer to
 the :doc:`../install/installation` section of this documentation.
 
 
@@ -29,6 +29,22 @@ This tutorial demonstrates a distributed version of the classic *Hello world* ex
 Basic usage of the library is described in this demo, just `follow the steps`_.
 
 .. _follow the steps: ../notebooks/hello_distributed_world.ipynb
+
+Automatic network generation
+============================
+
+In this tutorial, we demonstrate how to generate a network automatically. Access `the network generation notebook`_.
+
+.. _the network generation notebook: ../notebooks/network_generators.ipynb
+
+Algorithm benchmarking
+======================
+
+This tutorial demonstrates how to benchmark a distributed algorithm. Access `the benchmarking notebook`_.
+
+.. _the benchmarking notebook: ../notebooks/benchmarking.ipynb
+
+
 
 
 Build date: |today|
