+++ pydistsim/__init__.py
@@ -1,3 +1,7 @@
+"""
+Complete source for the pydistsim package.
+"""
+
 # flake8: noqa: F401
 
 from importlib import metadata
@@ -17,13 +21,10 @@ os.environ["QT_API"] = "pyside"
 # Declare namespace package
 from pkgutil import extend_path  # @Reimport
 
-from pydistsim.conf import settings
-from pydistsim.logger import logger, set_log_level
+from pydistsim.algorithm import NodeAlgorithm, StatusValues
+from pydistsim.benchmark import AlgorithmBenchmark
+from pydistsim.logger import disable_logger, enable_logger, logger
 from pydistsim.network import BidirectionalNetwork, Network, NetworkGenerator, Node
-from pydistsim.network.environment import Environment
-from pydistsim.npickle import *
-from pydistsim.sensor import CompositeSensor
 from pydistsim.simulation import Simulation
-from pydistsim.utils.localization import *
 
 __path__ = extend_path(__path__, __name__)  # @ReservedAssignment
