+++ pydistsim/_exceptions.py
@@ -1,3 +1,7 @@
+"""
+Custom exceptions for the PyDistSim package.
+"""
+
 from enum import StrEnum
 
 
@@ -8,21 +12,6 @@ class MessageUndeliverableException(Exception):
 
     def __str__(self):
         return self.e + repr(self.message)
-
-
-class NetworkErrorMsg(StrEnum):
-    NODE = "Node is already in another network."
-    NODE_SPACE = "Given position is not free space."
-    NODE_NOT_IN_NET = "Node not in network."
-
-
-class SimulationErrorMsg(StrEnum):
-    ALGORITHM = (
-        "Algorithms must be in tuple (AlgorithmClass,)"
-        " or in form: ((AlgorithmClass, params_dict),)."
-        "AlgorithmClass should be subclass of Algorithm"
-    )
-    ALGORITHM_NOT_FOUND = "Algorithm not found in network."
 
 
 class PyDistSimException(Exception):
@@ -43,7 +32,10 @@ class NetworkException(PyDistSimException):
     Exception class for network-related errors.
     """
 
-    ERRORS = NetworkErrorMsg
+    class ERRORS(StrEnum):
+        NODE = "Node is already in another network."
+        NODE_SPACE = "Given position is not free space."
+        NODE_NOT_IN_NET = "Node not in network."
 
 
 class SimulationException(PyDistSimException):
@@ -51,4 +43,10 @@ class SimulationException(PyDistSimException):
     Exception class for simulation-related errors.
     """
 
-    ERRORS = SimulationErrorMsg
+    class ERRORS(StrEnum):
+        ALGORITHM = (
+            "Algorithms must be in tuple (AlgorithmClass,)"
+            " or in form: ((AlgorithmClass, params_dict),)."
+            "AlgorithmClass should be subclass of Algorithm"
+        )
+        ALGORITHM_NOT_FOUND = "Algorithm not found in network."
