+++ pydistsim/demo_algorithms/__init__.py
@@ -1,3 +1,7 @@
+"""
+Example algorithms for test purposes and demonstrations.
+"""
+
 from pkgutil import extend_path
 
 __path__ = extend_path(__path__, __name__)  # @ReservedAssignment
