+++ pydistsim/gui/__init__.py
@@ -1,3 +1,7 @@
+"""
+GUI package for pydistsim.
+"""
+
 from .dictionarytreemodel import *
 from .simulationgui import *
 from .simulationui import *
