+++ pydistsim/network/__init__.py
@@ -1,10 +1,18 @@
+"""
+This module contains classes for representing networks and generating networks.
+"""
+
 # flake8: noqa: F401
 
-from ..simulation import AlgorithmsParam
-from .generator import NetworkGenerator, NetworkGeneratorException
-from .network import BidirectionalNetwork, Network, NetworkException, NetworkType
-from .node import Node
-from .rangenetwork import (
+from pydistsim.network.generator import NetworkGenerator, NetworkGeneratorException
+from pydistsim.network.network import (
+    BidirectionalNetwork,
+    Network,
+    NetworkException,
+    NetworkType,
+)
+from pydistsim.network.node import Node
+from pydistsim.network.rangenetwork import (
     BidirectionalRangeNetwork,
     CompleteRangeType,
     RangeNetwork,
@@ -13,3 +21,4 @@ from .rangenetwork import (
     SquareDiscRangeType,
     UdgRangeType,
 )
+from pydistsim.simulation import AlgorithmsParam
