+++ pydistsim/scripts/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module contains the runnable scripts of this package.
+"""
