+++ pydistsim/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Module for testing the PyDistSim package.
+"""
