+++ pydistsim/tests/test_algorithm_communicate.py
@@ -4,7 +4,7 @@ import unittest
 
 from pydistsim.message import Message
 from pydistsim.network import NetworkGenerator
-from pydistsim.network.behavior import ExampleProperties, behaviorModel
+from pydistsim.network.behavior import ExampleProperties, NetworkBehaviorModel
 
 CANT_MESSAGES = 100
 
@@ -59,7 +59,7 @@ def delay_only_first_message(network, message):
     return 100
 
 
-DelayOnlyFirstMessage = behaviorModel(
+DelayOnlyFirstMessage = NetworkBehaviorModel(
     message_ordering=True,
     message_delay_indicator=delay_only_first_message,
     bounded_communication_delays=True,
